@@ -1,3 +1,3 @@
 module Dacom
-  VERSION = "1.0.1"
+  VERSION = "2.0.0"
 end
